renamed files to make them consistent, added css config
