+++ src/pages/sign-up/sign-up.tsx
@@ -2,8 +2,8 @@ import { useAuthCreateUserWithEmailAndPassword } from "@react-query-firebase/aut
 import { auth } from "../../firebase/firebase";
 import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import css from "./signUp.module.scss";
-import { GoogleAuthButton } from "../../components/googleAuthButton/googleAuthButton";
+import css from "./sign-up.module.scss";
+import { GoogleAuthButton } from "../../components/google-auth-button/google-auth-button";
 
 export const SignUp = () => {
   const navigate = useNavigate();
